refactor: remove unused code, cleanup imports.
