+++ cmd/conduit/internal/initialize/init.go
@@ -11,11 +11,10 @@ import (
 
 	"github.com/spf13/cobra"
 
-	algodimporter "github.com/algorand/indexer/conduit/plugins/importers/algod"
-
-	"github.com/algorand/conduit/conduit"
+	"github.com/algorand/conduit/conduit/data"
 	"github.com/algorand/conduit/conduit/pipeline"
 	"github.com/algorand/conduit/conduit/plugins/exporters/filewriter"
+	algodimporter "github.com/algorand/conduit/conduit/plugins/importers/algod"
 )
 
 // InitCommand is the init subcommand.
@@ -58,7 +57,7 @@ func runConduitInit(path string, importerFlag string, processorsFlag []string, e
 		return err
 	}
 
-	configFilePath := filepath.Join(path, conduit.DefaultConfigName)
+	configFilePath := filepath.Join(path, data.DefaultConfigName)
 	f, err := os.Create(configFilePath)
 	if err != nil {
 		return fmt.Errorf("runConduitInit(): failed to create %s", configFilePath)
